+++ cms/src/features/posts/Posts.tsx
@@ -1,23 +1,19 @@
 import { useEffect, useState } from "react";
 import PostForm from "./components/PostForm";
-import Post from "./components/Post";
+
 import { useUser } from "../../hooks/useUser";
+import PostList from "./components/PostList";
 
 const Posts = () => {
-  // Get user's posts from database?
-
-  // const [posts, setPosts] = useState([]);
-  // const postList = posts.length > 0 && posts.map((post) => <Post />);
-
-  const { username } = useUser();
-  useEffect(() => {
-    console.log("username:", username);
-  }, [username]);
+  const { user } = useUser();
+  const data = user?.blog?.posts || [];
 
   return (
     <>
       <h2>Posts</h2>
+      <p>username {user?.username}</p>
       <PostForm />
+      <PostList data={data} />
     </>
   );
 };