+++ cms/src/features/posts/components/Post.tsx
@@ -1,5 +1,0 @@
-const Post = () => {
-  return <p>A post</p>;
-};
-
-export default Post;