+++ cms/src/hooks/useUser.tsx
@@ -7,11 +7,13 @@ import React, {
 } from "react";
 import { User } from "../interfaces/user";
 import { useAuth } from "./useAuth";
-import { axiosPrivate } from "../config/axios.config";
 import { AuthUserResponse } from "../interfaces/responses";
+import useAxiosPrivate from "./useAxiosPrivate";
+import { Post } from "../interfaces/blog";
 
 interface UserContext {
-  username?: string;
+  user?: User | null;
+  addPost: (newPost: Post) => void;
 }
 
 const UserContext = createContext<UserContext>({} as UserContext);
@@ -19,33 +21,49 @@ const UserContext = createContext<UserContext>({} as UserContext);
 const UserProvider: React.FC<{ children: React.ReactNode }> = ({
   children,
 }) => {
-  const { isAuthenticated } = useAuth();
+  const { accessToken, isAuthenticated } = useAuth();
   const [user, setUser] = useState<User | null>(null);
+  const axiosPrivate = useAxiosPrivate();
 
-  // const addPost = (post) => {};
+  const addPost = (newPost: Post) => {
+    setUser((prevUser) => {
+      if (!prevUser) return prevUser;
+      return {
+        ...prevUser,
+        ...(prevUser.blog && {
+          blog: { ...prevUser.blog, posts: [...prevUser.blog.posts, newPost] },
+        }),
+      };
+    });
+  };
 
   useEffect(() => {
     console.log("UserProvider mounted");
     const getUser = async () => {
-      try {
-        const response = await axiosPrivate.get<AuthUserResponse>("/auth/user");
-        setUser(response.data.user);
-        console.log("response:", response);
-      } catch (err) {
-        console.error(err);
+      if (isAuthenticated) {
+        try {
+          const response = await axiosPrivate.get<AuthUserResponse>(
+            "/auth/user"
+          );
+          setUser(response.data.user);
+          console.log("response:", response);
+        } catch (err) {
+          console.error(err);
+        }
       }
     };
 
-    if (isAuthenticated) {
+    console.log("accessToken in UserProvider:", accessToken);
+    if (accessToken) {
       // Send GET request to API to get user profile
       getUser();
+    } else {
+      setUser(null);
     }
   }, [isAuthenticated]);
 
   const useUserValue = useMemo(() => {
-    return {
-      username: user?.username,
-    };
+    return { user, addPost };
   }, [user]);
 
   return (