+++ cms/src/interfaces/responses.ts
@@ -1,10 +1,22 @@
 import { User } from "./user";
+import { Post } from "./blog";
 
-export interface AuthSuccessResponse<T = void> {
+export interface SuccessResponse {
   code: 200;
   status: "success";
+}
+
+export interface AuthSuccessResponse extends SuccessResponse {
   accessToken: string;
-  user: T;
+}
+
+export interface AuthUserResponse extends AuthSuccessResponse {
+  user: User;
+}
+
+// A new post has been created
+export interface PostSuccessResponse extends SuccessResponse {
+  post: Post;
 }
 
 export interface LoginErrorResponse {